+++ CHANGELOG.md
@@ -36,6 +36,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Snippet insertion inserts first node correctly
 - Enable no-bare-strings rule and update translations where needed
 
+### Changed
+- Use one-way-binding in variable label input
+- Use one-way-binding in number variable inputs
+
 ### Added
 - Addition of `variable` group to each variable node-spec.